replace lightbox library
add backdrop style
