+++ src/partials/modal-details.html
@@ -1,5 +1,5 @@
 <!--Модальное окно про фильм-->
-<div class="backdrop is-hidden">
+<div class="backdrop backdrop__movie-info is-hidden">
   <!--< div class="modal">
     <button class="modal-close-btn">
       <svg class="close-icon" width="14" height="14">
